requirements.txt
PyQt6>=6.4.0
pyyaml>=6.0
crewai>=0.1.0
crewai-tools>=0.1.0